First commit
Trying out how easy it would be to create a gujarati letter classifier using an out of the box methodology. followed a tutorial for a refresher and got a pretty high accuracy, but I want to inspect the model more and check to see if it actually does what it says it does. After that, I will test a few model configs and try to make a simple Gradio based web app!

I will eventually want this to learn handwriting, so it would be good to stop and try to think on how to get it to learn this manually. I'm planning to have some sort of a learning procedure, where I use the tool for some period of time, and try to manually create a dataset.